LayoutTests/fast/eventloop/break-all-to-allow-rendering-update-expected.txt

This tests breakToAllowRenderingUpdate successfully stops the current run of the event loop.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS logs.join(", ") is "1, 2, 3, 4"
PASS successfullyParsed is true

TEST COMPLETE
